f17-Added groups to the command modules in debug mode to more easily distinguish between the commands of the debug and the release bot. This required breaking down Random into two modules.
